+++ blacksheep/baseapp.pyx
@@ -6,6 +6,7 @@ from .exceptions cimport HTTPException, NotFound
 import os
 import html
 import http
+import logging
 import traceback
 
 
@@ -21,18 +22,42 @@ async def common_http_exception_handler(app, Request request, HTTPException http
     return Response(http_exception.status, content=TextContent(http.HTTPStatus(http_exception.status).phrase))
 
 
+def get_logger():
+    logger = logging.getLogger("blacksheep.server")
+    logger.setLevel(logging.INFO)
+    return logger
+
+
 cdef class BaseApplication:
 
     def __init__(self, bint show_error_details, object router):
         self.router = router
         self.exceptions_handlers = self.init_exceptions_handlers()
         self.show_error_details = show_error_details
+        self.logger = get_logger()
 
     def init_exceptions_handlers(self):
         return {
             404: handle_not_found,
             400: handle_bad_request
         }
+
+    async def log_unhandled_exc(self, request, exc):
+        self.logger.error(
+            "Unhandled exception - \"%s %s\"",
+            request.method,
+            request.url.value.decode(),
+            exc_info=exc
+        )
+
+    async def log_handled_exc(self, request, HTTPException exc):
+        self.logger.info(
+            "HTTP %s - \"%s %s\". %s",
+            exc.status,
+            request.method,
+            request.url.value.decode(),
+            str(exc)
+        )
 
     async def handle(self, Request request):
         cdef object route
@@ -46,8 +71,10 @@ cdef class BaseApplication:
             try:
                 response = await route.handler(request)
             except HTTPException as http_exception:
+                await self.log_handled_exc(request, http_exception)
                 response = await self.handle_http_exception(request, http_exception)
             except Exception as exc:
+                await self.log_unhandled_exc(request, exc)
                 response = await self.handle_exception(request, exc)
         else:
             response = await self.exceptions_handlers.get(404)(self, request, None)
@@ -67,20 +94,27 @@ cdef class BaseApplication:
 
     async def handle_internal_server_error(self, Request request, Exception exc):
         if self.show_error_details:
-            tb = traceback.format_exception(exc.__class__,
-                                            exc,
-                                            exc.__traceback__)
+            tb = traceback.format_exception(
+                exc.__class__,
+                exc,
+                exc.__traceback__
+            )
             info = ''
             for item in tb:
                 info += f'<li><pre>{html.escape(item)}</pre></li>'
 
-            content = HtmlContent(self.resources.error_page_html
-                                  .format_map({'process_id': os.getpid(),
-                                               'info': info,
-                                               'exctype': exc.__class__.__name__,
-                                               'excmessage': str(exc),
-                                               'method': request.method,
-                                               'path': request.url.value.decode()}))
+            content = HtmlContent(
+                self.resources.error_page_html.format_map(
+                    {
+                        'process_id': os.getpid(),
+                        'info': info,
+                        'exctype': exc.__class__.__name__,
+                        'excmessage': str(exc),
+                        'method': request.method,
+                        'path': request.url.value.decode()
+                    }
+                )
+            )
 
             return Response(500, content=content)
         return Response(500, content=TextContent('Internal server error.'))