+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2021-01-24 :grapes:
+- Logs handled and unhandled exceptions (fixes: #75)
+- Adds support for [Flask Variable Rules syntax](https://flask.palletsprojects.com/en/1.1.x/quickstart/?highlight=routing#variable-rules) (ref. #76) and more granular control on the
+  route parameters' patterns when matching web requests
+- Adds the missing `html` method to the `Controller` class (#77) - thanks to
+  [skivis](https://github.com/RobertoPrevato/BlackSheep/commits?author=skivis)!
+- Deprecates the `ServeFilesOptions` class and reduces verbosity of the
+  `Application.serve_files` method (#71)
+
 ## [0.3.1] - 2020-12-27 🎄
 - Implements an abstraction layer to [handle CORS](https://www.neoteroi.dev/blacksheep/cors/)
 - Improves the code API to handle [response cookies](https://www.neoteroi.dev/blacksheep/responses/#setting-cookies)